Fixing size

+++ src/testPkt_rx.c
@@ -86,7 +86,7 @@ int main() {
 	uint8_t rx_pktno = 0;
 	uint16_t rx_pktsz = 0;
 	uint16_t pkt_sz = DATA_SZ;
-	uint8_t magic_no = 0;
+	uint16_t magic_no = 0;
 
 	// pcap
 	const u_char *packet;
@@ -105,7 +105,7 @@ int main() {
 	uint8_t *pktbuf;
 	uint8_t *pktnobuf;
 	uint16_t *pktszbuf;
-	uint8_t *pktmagicbuf;
+	uint16_t *pktmagicbuf;
 
 	// GPIO mraa
 	mraa_result_t ret = MRAA_SUCCESS;
@@ -227,8 +227,8 @@ int main() {
 			memcpy(&rx_pktno, pktnobuf, sizeof(uint8_t));
 			pktszbuf = (uint16_t *) (pktbuf + 97); // 97 --> magic location
 			memcpy(&rx_pktsz, pktszbuf, sizeof(uint16_t));
-			pktmagicbuf = (uint8_t *) (pktbuf + 98); // 98 --> magic location
-			memcpy(&magic_no, pktmagicbuf, sizeof(uint8_t));
+			pktmagicbuf = (uint16_t *) (pktbuf + 98); // 98 --> magic location
+			memcpy(&magic_no, pktmagicbuf, sizeof(uint16_t));
 
 #ifdef DEBUG
 			// Get pkt length
